% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA.R
\name{apa}
\alias{apa}
\title{apa: Turn statistical test output into APA style citations.
This takes a number of test types and outputs an APA style output for the test.
IE \eqn{Statistic Name_(Degrees of Freedom)=.xx CI[.xx, .xx], p<.xx} \cr
If latex is selected, the inline statement must be wrapped in \code{$} or \code{$$} to generate latex properly.}
\usage{
apa(tout, type = "html")
}
\arguments{
\item{tout}{The output list from a statistical test. See \bold{Details} for supported test types.}

\item{type}{\code{(character)} The type of APA style output required, html or latex, defaults to html.}
}
\value{
Either an html widget or latex with the APA style output.
This function is useful when writing up results from statistical analyses to provide inline APA style declaration of the
results given the resulting output of the function.
See \href{https://rmarkdown.rstudio.com/lesson-4.html}{knitr documentation} for info on inline r code.
Currently compatible test types are:
\describe{
 \item{\code{kruskal.test}}{Kruskal-Wallis rank sum test}
 \item{\code{wilcox.text}}{Wilcoxon Rank-sum (Mann-Whitney) test}
 \item{\code{t.test}}{Difference in means test}
 \item{\code{anova | aov}}{Analysis of Variance. \emph{Note:} Multiple variables will be return as a named list. The APA output for the specific variable can be returned by calling the index by the name of the variable}
 \item{\code{effsize::cohen.d}}{Cohen's D and Hedge's G effect sizes}
 \item{\code{shapiro.test}}{Shapiro-Wilk test of normality}
 \item{\code{car::leveneTest}}{Levene's test for homogeneity of variance across groups}
 \item{\code{bartlett.test}}{Bartlett Test of Homogeneity of Variances}
 \item{\code{fligner.test}}{Fligner-Killeen Test of Homogeneity of Variances}
 \item{\code{cor(method="pearson")}}{Pearson's (r) product-moment test of correlation}
}
}
\description{
apa: Turn statistical test output into APA style citations.
This takes a number of test types and outputs an APA style output for the test.
IE \eqn{Statistic Name_(Degrees of Freedom)=.xx CI[.xx, .xx], p<.xx} \cr
If latex is selected, the inline statement must be wrapped in \code{$} or \code{$$} to generate latex properly.
}
\examples{
dat <- data.frame(y = rnorm(15,0,1),x = {rnorm(15,0,1) + rnorm(15,0,.02)})
apa(t.test(y = dat$y, x = dat$x))
apa(t.test(y = dat$y, x = dat$x), type = "latex")
}
