% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractHighlights.R
\name{extractHighlights}
\alias{extractHighlights}
\title{extractHighlights}
\usage{
extractHighlights(document, cit = F)
}
\arguments{
\item{document}{\code{(character)} The path to and the name of the filename ie. "my directory/myfile.pdf"}

\item{cit}{\code{(boolean or character)} If highlights come from a PDF with metadata, set this to TRUE and each of the highlights will be accompanied by an APA style citation using the pdf metadata. If citation is not formatted properly, this can be set as the character string of the APA style citation which will be appended to each highlight. Defaults to FALSE.}
}
\value{
Parsed highlights that are formatted correctly will display properly when knitted to an HTML document and chunk options are set to display results='asis'
}
\description{
Easily extract highlights made with a \href{http://www.docear.org/support/user-manual/#pdf_editors}{compatible PDF reader} from PDF documents, combine multi-page highlights, and apply formatting based on specific markup.
}
\details{
Highlighting markup is as follows:
\itemize{
 \item{\code{*Text*}}{ Will italicize alphanumeric text as per RMarkdown markup - also works when knitting to PDF}
 \item{\code{i\{text\}}}{ Will remove brackets and italicize and make light blue all alphanumeric text including only the following symbols: ,.&-:_ Works when knitting to all formats}
 \item{\code{**Text**}}{ Will bold all alphanumeric text as per RMarkdown markup - also works when knitting to all formats}
 \item{\code{b\{text\}}}{ Will remove brackets and bold all alphanumeric text including only the following symbols: ,.&-:_ Works when knitting to HTML only}
 \item{\code{\#Keyword}}{ Will make text bold retaining the hash for easy identification. Allows tagging of highlights. Any bracketed text that immediately follows a keyword with no space (\{Text\}) will be italicized and made light blue. Allows for description of the highlight or rationale for the keyword.}
 \item{\code{color\{\}}}{ Will make text the \href{https://www.w3schools.com/colors/colors_names.asp}{named HTML color}. Can be combined with asterisk-based italicization and bolding.}
 \item{\code{|color}}{A named HTML color following the pipe at the beginning of the highlight will create a border box around the text with the left border as the color specified. Useful for color-coding and marking important highlights.}
}
}
