% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{find_peaks}
\alias{find_peaks}
\title{find_peaks}
\usage{
find_peaks(x, m = 3)
}
\arguments{
\item{x}{\code{(numeric)} A vector object}

\item{m}{\code{(numeric)} A steepness of slope numeric as a threshold for calculating peaks. The larger the number the fewer the peaks meeting the criteria.}
}
\value{
\item{(numeric)}{ A numeric vector of exact indices of peaks.}
\emph{Note:} this differs from \code{\link[quantmod:findPeaks]{quantmod::findPeaks}} in that is supplies the indices of the exact peaks, whereas quantmod supplies the vector of indices lagged by one time unit - useful when selling a stock but not if you want to know the exact peak locations.
}
\description{
Find Peaks in a timeseries. Credit: \url{https://github.com/stas-g}
}
\examples{
dat <- data.frame(y = rnorm(15,0,1),x = {rnorm(15,0,1) + rnorm(15,0,.02)})
apply(dat, 2, find_peaks)
}
