% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{ggColor}
\alias{ggColor}
\title{ggColor}
\usage{
ggColor(n)
}
\arguments{
\item{n}{\code{(integer)} The number of colors from the default color wheel you would like to return}
}
\value{
\code{(character)} Character vector of hexadecimal color values
}
\description{
Generate hexadecimal values from R's default color wheel.
Credit: \url{https://stackoverflow.com/users/412342/john-colby}
}
\examples{
temp <- data.frame(num = c(5, 7, 6, 4, 8), fac = factor(c("a", "a", "b", "b", "b")))
ggplot2::qplot(data = temp, x = num, fill = fac)
par(mfrow=c(1,2))
cls <- ggColor(2)
barplot(temp$num, col = cls[1])
barplot(temp$num, col = cls[2])
}
