% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{\%n\%}
\alias{\%n\%}
\title{grapes n grapes tests if a value is length 0, NULL, or NA and returns FALSE if it is, otherwise returns TRUE.}
\usage{
x \%n\% y
}
\arguments{
\item{x}{Any object. Accepts expressions enclosed in {}}

\item{y}{Anything, not used in the calculation}
}
\value{
\code{(boolean)} A TRUE indicates the object is not NULL, NA or length 0.
}
\description{
Useful for conditional evaluation of a valid value for an object for further calculations or functions.
}
\examples{
{4/0} \%n\% T
# Infinite is an acceptable value
NaN \%n\% T
NA \%n\% T
NULL \%n\% T
numeric(0) \%n\% T
}
