% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssumptionEvaluation.R
\name{homoVariance}
\alias{homoVariance}
\title{homoVariance}
\usage{
homoVariance(form, dat)
}
\arguments{
\item{form}{\code{(formula)} A formula in the form amenable to the lm function for which homogeneity of variance between predictors will be tested.}

\item{dat}{\code{(data.frame)} The data in data.frame format}
}
\value{
A list with the output of each test
   \describe{
     \item{Levene}{Levene's test for homogeneity of variance across groups, see \link[car]{leveneTest}}
     \item{Bartlett}{Bartlett's test for homogeneity of variance, see \link{bartlett.test}}
     \item{Fligner}{Fligner-Killeen test for homogeneity of variance, see \link{fligner.test}}
   }
}
\description{
Performs the three most common tests of homogeneity of variance.
This test performs three of the most common homogeneity of variance tests:
\itemize{
     \item{\code{\link[car:leveneTest]{leveneTest}}}
     \item{\code{\link[stats:bartlett.test]{bartlett.test}}}
     \item{\code{\link[stats:fligner.test]{fligner.test}}}
}
}
\examples{
dat <- data.frame(y = rnorm(15,0,1),x = {rnorm(15,0,1) + rnorm(15,0,.02)})
homoVariance(y ~ x, dat)
}
