% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{Mode}
\alias{Mode}
\title{Mode}
\usage{
Mode(v)
}
\arguments{
\item{v}{\code{(numeric)} a numeric vector}
}
\value{
\code{(numeric)} The mode as a numeric
}
\description{
Computes the mode of a numeric vector. R does not have a built in function for computing the mode (the most frequent value) in a vector of numeric values. This function (\emph{Note: Mode with a capital M}) returns the mode.
}
\examples{
x <- c(rep(1:3, each = 2),rep(4,3))
Mode(x)
dat <- data.frame(y = rnorm(15,0,1),x = {rnorm(15,0,1) + rnorm(15,0,.02)})
apply(dat, 2, Mode)
}
