% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA.R
\name{p.txt}
\alias{p.txt}
\title{p.txt formats a p-value as an APA style character vector.}
\usage{
p.txt(p.val)
}
\arguments{
\item{p.val}{\code{(numeric)} A numeric p-value (or vector thereof)}
}
\value{
\item{p text}{The p-value as an APA style character vector}
Useful for inline r calls within the text of an Rmd.
See \href{https://rmarkdown.rstudio.com/lesson-4.html}{knitr documentation} for info on inline r code.
}
\description{
A p-value of .07 will return 'p<.1', a p-value of .04: P<.05 etc.
}
\details{
If a p-value is not significant (p>.1) the p-value will be reported to 2 decimal places.
}
\examples{
p.txt(.07)
# "p<.1"
p.txt(.04)
# "p<.05"
}
