% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{rleIndex}
\alias{rleIndex}
\title{rleIndex}
\usage{
rleIndex(input_rle)
}
\arguments{
\item{x}{\code{(rle)} An rle object}
}
\value{
\item{(data.frame)}{ A data.frame with length, values, start and end indices.}
}
\description{
Given an \code{\link[rle]{rle}} this function will return a dataframe of starts, ends, and indexes thereof of the run lengths.
Credit: \url{https://stackoverflow.com/questions/43875716/find-start-and-end-positions-indices-of-runs-consecutive-values}
}
\examples{
dat <- data.frame(y = rnorm(15,0,1),x = {rnorm(15,0,1) + rnorm(15,0,.02)})
rleIndex(rle(abs(dat$x) > .5))
}
