% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{startPkgs}
\alias{startPkgs}
\title{startPkgs}
\usage{
startPkgs(pkgs)
}
\arguments{
\item{pkgs}{\code{(character)} Vector of package names to be loaded in the R environment}
}
\description{
Loads all packages supplied by the character vector silently and quickly. Useful for invisibly starting all library requirements with less typing in the setup chunk rather than calling each library independently.
}
\examples{
req.pkgs <- c("tidyverse","magrittr","dplyr")
system.time({library(tidyverse);library(magrittr);library(dplyr)})
system.time(startPkgs(req.pkgs))
}
