% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssumptionEvaluation.R
\name{testTrans}
\alias{testTrans}
\title{testTrans}
\usage{
testTrans(dvs, ivs, dat)
}
\arguments{
\item{dvs}{\code{(character)} The names of the dependent variables as a character vector}

\item{ivs}{\code{(character)} The name of the indpendent variables as a character vector}

\item{dat}{\code{(data.frame)} The data.frame that contains the variables}
}
\value{
\item{list}{ of ggplot2 grobs showing the relationship of each variable to each other variable}
Of note, the declaration of the best-fitting transformations for each relationship can be accessed via \code{lapply(result, purrr::pluck,"labels","caption")}
}
\description{
Test mathematical transformations (exp, log, square, cube) of predictor variables.
For a set of dependent variables and independent variables testTrans will test the logarithmic, exponential, quadratic, and cubic relationships of the DV between each DV & IV and provide visual feedback to allow determination of the best fitting mathematical tranformation of each dependent variable for a given independent variable.
}
\examples{
dat <- data.frame(y = rnorm(15,0,1),x = {rnorm(15,0,1) + rnorm(15,0,.02)})
testTrans("y", "x", dat)
}
