% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{toNum}
\alias{toNum}
\title{toNum}
\usage{
toNum(x)
}
\arguments{
\item{x}{\code{(character)} A character vector in the form "$12.23"}
}
\value{
\code{(numeric)} A numeric vector
}
\description{
A simple function to convert financials retrieved from googlesheets as characters to numeric
}
