% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{unloadPkgs}
\alias{unloadPkgs}
\title{unloadPkgs silently unloads all packages in the supplied character vector}
\usage{
unloadPkgs(pkgs)
}
\arguments{
\item{pkgs}{\code{(character)} Vector of package names to be unloaded from the R environment}
}
\description{
Large packages loaded into the R environment can slow down computation. Unload those packages easily and silently with unloadPkgs
}
\examples{
startPkgs(c("tidyverse","magrittr"))
system.time({detach("package:tidyverse", character.only = TRUE)
detach("package:magrittr", character.only = TRUE)})
startPkgs(c("tidyverse","magrittr"))
system.time(unloadPkgs(c("tidyverse","magrittr")))
}
