% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssumptionEvaluation.R
\name{visEDA}
\alias{visEDA}
\title{visEDA}
\usage{
visEDA(dat)
}
\arguments{
\item{dat}{\code{(data.frame)} A data.frame for which variables are to be explored.}
}
\value{
A \code{gridExtra::grid.arrange} grob object with a Density plot, Normal QQ plot and Histogram for each numeric variable in the data.frame. Each \code{grid.arrange} grob contains the following graphs:
 /describe{
   \item{\code{a}}{ The first graph is a density plot with the mean and standard deviations marked. The caption contains the results of the Shapiro-Wilks test of Normality.}
   \item{\code{qq}}{ The second graph is a QQ plot to evaluate normality of the distribution.}
   \item{\code{b}}{ The third graph is a histogram with density curve plotted. The caption contains the results of the two-sided Kolmogorov-Smirnov test of Normality}
}
}
\description{
A function for visual exploratory data analysis.
This function provides three common plots useful for visually exploring data and testing assumptions prior to statistical analysis.
}
\examples{
dat <- data.frame(y = rnorm(15,0,1),x = {rnorm(15,0,1) + rnorm(15,0,.02)})
visEDA(dat)
}
\seealso{
\code{\link[stats:shapiro.test]{shapiro.test}}, \code{\link[stats:ks.test]{ks.test}}, \code{\link[gridExtra:arrangeGrob]{grid.arrange}}, \code{\link[ggpubr:ggqqplot]{ggqqplot}}
}
