% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{findna}
\alias{findna}
\title{findna: Report summary of NA values and their Indices}
\usage{
findna(dat)
}
\arguments{
\item{dat}{\code{(data.frame, matrix)} data.frame, vector or matrix object}
}
\value{
\item{list} object with the following values:
\describe{
 \item{`NA`}{ A vector}
   \itemize{
     \item{\code{No.}}{The Number of NA}
     \item{\code{P}}{ The percentage as a decimal of \eqn{\frac{\text{the number of NA values}}{\text{the total number of values in the column}}}.}
   }
 \item{\code{Indices}}{ The indices of the row numbers that contain NA in that column, useful for locating NA and examining adjacent data.}}
}
\description{
Iterates over the columns of a data.frame or matrix and returns descriptive summary of NA values
}
\examples{
findna(c(1, NA, 2))
dat <- data.frame(y = rnorm(15,0,1),x = {rnorm(15,0,1) + rnorm(15,0,.02)})
dat[sample(1:nrow(dat), size = 3), "x"] <- NA
findna(dat)
}
